Moving Hero, Main and Team to sections/ folder
