+++ src/components/views/Home/Home.tsx
@@ -1,9 +1,9 @@
 import './Home.styles.scss'
 
 import { Header } from '../../Header'
-import { HeroSection } from './HeroSection'
-import { MainSection } from './MainSection'
-import { TeamSection } from './TeamSection'
+import { HeroSection } from './sections/HeroSection'
+import { MainSection } from './sections/MainSection'
+import { TeamSection } from './sections/TeamSection'
 
 export const Home = () => {
   return (
